# docs/api-specification.yaml
openapi: 3.0.0
info:
  title: IdeaBox API
  description: |
    API documentation for the IdeaBox social media platform.

    **Authentication:** Most endpoints require a valid JWT token obtained from `/api/auth/signin`.
    Include the token in the `Authorization` header as: `Bearer YOUR_TOKEN`

    **For Backend Developers:**
    - Each endpoint spec below shows what you need to build
    - Follow the example endpoints in `app/api/posts/route.ts`
    - Use Prisma to interact with the database
    - Always validate user input before saving to database
  version: 1.0.0
  contact:
    name: IdeaBox Team
servers:
  - url: http://localhost:3000
    description: Local development server

# Reusable components
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from NextAuth signin
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      example:
        error: "Something went wrong"
...
    Post:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        userId:
          type: string
        projectId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: "clx1234567890"
        content: "Just launched my new app!"
        imageUrl: "https://example.com/screenshot.png"
        userId: "clx0987654321"
        projectId: null
        createdAt: "2024-01-01T00:00:00.000Z"
        updatedAt: "2024-01-01T00:00:00.000Z"
...
security:
  - BearerAuth: []
paths:
...
  /api/posts:
    get:
      summary: Get all posts
      description: Retrieve a feed of all posts (newest first)
      tags:
        - Posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

    post:
      summary: Create a new post
      description: Create a new post/idea
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "Check out my new project idea!"
                imageUrl:
                  type: string
                  example: "https://example.com/image.jpg"
                projectId:
                  type: string
                  example: "clx1234567890"
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/posts/{id}:
    get:
      summary: Get a post by ID
      description: Retrieve a single post with its details
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a post
      description: Update a post (only the author can update)
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                imageUrl:
                  type: string
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the post author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a post
      description: Delete a post (only the author can delete)
      tags:
        - Posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post deleted successfully"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  # ==================== COMMENT ENDPOINTS ====================
...